requirements.txt
yt-dlp
faster-whisper
openai
python-dotenv
youtube-transcript-api
ollama
torch
python-dateutil
